New project "testing-common"
The "cards" and "engine-integration-tests" projects relied on the JUnitTest class to run tests. Therefore, that class had to live in "src/main/java" of some project, but not be distributed with "jmagic.jar". The only projects we don't distribute are "development" and "engine-integration-tests" and putting JUnitTest in either of those would cause a circular dependency, so a new project was necessary.